move rollup to build package
